(build): the one translation unit that compiles the front door had no warning flags
CMakeLists.txt says why idemip_warnings is an INTERFACE target rather than PRIVATE to the library:
"Attached PRIVATE to the library it never compiled most of the tree: no .c in src/ includes tcp.h or
ipv6.h or udp.h, so those headers were built only by a test TU, which had no warning flags at all."
Three targets were still linking idemip without it, and one of them is load-bearing.

3ee0068 pointed tools/idemip_sizes.c at src/idemip.h precisely because it is the only target that
builds in all thirteen capability sets, so the front door reaches a compiler in every one of them
rather than in the two that build test_loopback. That target links idemip and nothing else, so the
whole front door - and the thirteen headers no .c in src/ includes - was being compiled with -Wall
off, -Wextra off, -Wconversion off, the lot. The one TU that exists to prove idemip.h compiles
everywhere was proving it at the compiler's default warning level, which is none.

The two benches were the same, for the same reason: they link idemip and take its optimisation
settings, and were reading the same headers with no warning set.

All three link idemip_warnings now. Nothing in src/ warns under it - -Wall -Wextra -Wpedantic
-Wshadow -Wconversion -Wsign-conversion -Wcast-align -Wundef over the front door's whole include
closure comes out silent - which is the answer the flags were meant to be asking for and were not.

What did warn was the tool's own printf. "%zu" is C99 and correct where CI builds, and gcc targeting
mingw reports "unknown conversion type character 'z'" for it, because msvcrt's printf never took the
length modifier; the tool prints unsigned long now, which every runtime takes and which no borrow in
this tree comes within three orders of magnitude of overflowing. IDEMIP_TOTAL_BORROW still reads
31624 and every row is unchanged.

All thirteen capability sets build clean and pass: all 56, no-udp 50, no-tcp 49, no-transport 45,
ip6-only 34, ip4-only 33, ip6-tcp 31, ip6-udp 31, ip4-tcp 29, ip6-bare 28, ip4-udp 28, ip4-bare 25,
eth-only 10. The benches build clean with IDEMIP_BENCH=ON.

+++ tools/idemip_sizes.c
@@ -24,7 +24,7 @@
 
 #include <stdio.h>
 
-#define ROW(name) printf("  %-28s %7zu\n", #name, (size_t)(name))
+#define ROW(name) printf("  %-28s %7lu\n", #name, (unsigned long)(name))
 
 int main(void)
 {
@@ -63,7 +63,7 @@ int main(void)
     ROW(IDEMIP_STATS_BORROW);
     ROW(IDEMIP_VLAN_BORROW);
     ROW(IDEMIP_ETHIP6_BORROW);
-    printf("  %-28s %7zu\n\n", "= IDEMIP_SHARED_BORROW", (size_t)IDEMIP_SHARED_BORROW);
+    printf("  %-28s %7lu\n\n", "= IDEMIP_SHARED_BORROW", (unsigned long)IDEMIP_SHARED_BORROW);
 
     printf("per interface, taken IDEMIP_NETIF_COUNT times:\n");
     ROW(IDEMIP_PHY_BORROW);
@@ -77,13 +77,13 @@ int main(void)
     ROW(IDEMIP_DAD_BORROW);
     ROW(IDEMIP_SLAAC_BORROW);
     ROW(IDEMIP_RDNSS_BORROW);
-    printf("  %-28s %7zu  x %u\n\n", "= IDEMIP_PER_NETIF_BORROW",
-           (size_t)IDEMIP_PER_NETIF_BORROW, (unsigned)IDEMIP_NETIF_COUNT);
+    printf("  %-28s %7lu  x %u\n\n", "= IDEMIP_PER_NETIF_BORROW",
+           (unsigned long)IDEMIP_PER_NETIF_BORROW, (unsigned)IDEMIP_NETIF_COUNT);
 
-    printf("  %-28s %7zu\n", "IDEMIP_TOTAL_BORROW", (size_t)IDEMIP_TOTAL_BORROW);
+    printf("  %-28s %7lu\n", "IDEMIP_TOTAL_BORROW", (unsigned long)IDEMIP_TOTAL_BORROW);
 
-    printf("\nnot counted: the driver's frame buffers (%u octets each, stride %zu),\n",
-           (unsigned)IDEMIP_DMA_FRAME_MAX, (size_t)IDEMIP_DMA_BUF_STRIDE);
+    printf("\nnot counted: the driver's frame buffers (%u octets each, stride %lu),\n",
+           (unsigned)IDEMIP_DMA_FRAME_MAX, (unsigned long)IDEMIP_DMA_BUF_STRIDE);
     printf("             and the caller's stack.\n");
     return 0;
 }
